refactor: move shared dropdown logic to combo-box items mixin
Combo-box and multi-select-combo-box kept identical copies of
requestContentUpdate, the clear-button content refresh and the
focused-index preservation in _setDropdownItems. Multi-select-combo-box
and time-picker also overrode _closeOrCommit with the base body.
