+++ packages/combo-box/src/vaadin-combo-box-items-mixin.d.ts
@@ -85,4 +85,12 @@ export declare class ComboBoxItemsMixinClass<TItem> {
    * @attr {none|first-match|only-match} partial-match-mode
    */
   partialMatchMode: ComboBoxPartialMatchMode;
+
+  /**
+   * Requests an update for the content of items.
+   * While performing the update, it invokes the renderer (passed in the `renderer` property) once an item.
+   *
+   * It is not guaranteed that the update happens immediately (synchronously) after it is requested.
+   */
+  requestContentUpdate(): void;
 }
