+++ packages/combo-box/src/vaadin-combo-box-mixin.d.ts
@@ -87,12 +87,4 @@ export declare class ComboBoxMixinClass<TItem> {
    * @attr {string} item-id-path
    */
   itemIdPath: string | null | undefined;
-
-  /**
-   * Requests an update for the content of items.
-   * While performing the update, it invokes the renderer (passed in the `renderer` property) once an item.
-   *
-   * It is not guaranteed that the update happens immediately (synchronously) after it is requested.
-   */
-  requestContentUpdate(): void;
 }
